sql_mcp_server/requirement.txt
langchain-core
langchain-google-genai
mcp
psycopg2-binary
requests
